默认数据库表引擎

+++ database/migrations/2019_10_15_025137_create_seckill_products_table.php
@@ -16,7 +16,8 @@ class CreateSeckillProductsTable extends Migration
         Schema::create('seckill_products', function (Blueprint $table) {
             $table->engine='InnoDB';
             $table->bigIncrements('id');
-            $table->unsignedBigInteger('product_id');
+            //$table->unsignedBigInteger('product_id');
+            $table->unsignedInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->dateTime('start_at');
             $table->dateTime('end_at');
